requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.29.0
idna==3.3
msgpack==1.0.4
py-algorand-sdk==1.16.1
pycparser==2.21
pycryptodomex==3.15.0
PyNaCl==1.5.0
python-decouple==3.6
requests==2.28.1
six==1.16.0
urllib3==1.26.12
